pyproject.toml
[tool.poetry]
name = "triton-inference"
version = "0.1.0"
description = ""
authors = ["Dmitry <kachkindmitry@gmail.com>"]
readme = "README.md"
packages = [{ include = "triton_server" }]

[tool.poetry.dependencies]
python = "^3.10"
nvidia-pytriton = "^0.5.4"
onnx = "^1.16.1"
onnxruntime = "^1.18.0"
pillow = "^10.3.0"
ecs-logging = "^2.1.0"
pydantic = "^2.7.3"
pydantic-settings = "^2.3.1"
boto3 = "^1.34.122"
numpy = "^1.26.4"

[tool.poetry.group.gpu.dependencies]
onnxruntime-gpu = "^1.18.0"

[tool.poetry.group.dev.dependencies]
torch = "^2.3.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pylint = "^3.2.3"
black = "^24.4.2"
mypy = "^1.10.0"
isort = "^5.13.2"
flake8 = "^7.0.0"



[tool.black]
 line-length = 79
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.tox
  | venv
  | build
  | dist
)/
'''


[tool.isort]
profile = "black"
multi_line_output = 3


[tool.mypy]
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
warn_unreachable = true
pretty = true

[[tool.mypy.overrides]]
module = "*.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.pylint.main]
fail-under = 10
ignore = ["data"]
ignored-modules = ["test"]
py-version = "3.10"
suggestion-mode = true

[tool.pylint.basic]
bad-names = ["foo", "bar", "baz", "toto", "tutu", "tata"]
class-attribute-naming-style = "any"
class-const-naming-style = "UPPER_CASE"
class-naming-style = "PascalCase"
const-naming-style = "UPPER_CASE"
function-naming-style = "snake_case"
property-classes = ["abc.abstractproperty"]


[tool.pylint.format]
ignore-long-lines = "^\\s*(# )?<?https?://\\S+>?$"
indent-after-paren = 4
indent-string = "    "
max-line-length = 79
max-module-lines = 1000


[tool.pylint.messages_control]
disable = [
    "missing-docstring",
    "too-few-public-methods",
    "too-many-arguments",
    "invalid-all-format",
    "import-error",
    "logging-fstring-interpolation",
    "broad-exception-caught"
]


good-names = []



[tool.pylint.similarities]
ignore-imports = true
ignore-docstrings = true
ignore-comments = true


[tool.pytest.ini_options]
asyncio_mode = "strict"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(levelname)-5.5s [%(name)s] %(message)s"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
